add public type in json/annotation.

See merge request !14

+++ searchentity/src/main/java/cn/superid/search/entities/user/file/FileSearchVO.java
@@ -17,7 +17,7 @@ public class FileSearchVO implements UserBasedIndex {
   public FileSearchVO() {
   }
 
-  public FileSearchVO(String id, Integer type) {
+  public FileSearchVO(String id, Byte type) {
     this.id = id;
     this.type = FileType.values()[type];
   }
