+++ searchlogic/src/main/java/cn/superid/search/impl/entities/user/user/UserPO.java
@@ -30,7 +30,7 @@ public class UserPO {
   private String superId;
   @Field(type = FieldType.Nested)
   private List<TagPO> tags;
-  @Field(type = FieldType.Integer)
+  @Field(type = FieldType.Byte)
   private Integer publicType;
 
   public UserPO() {
